test(sessioncmd): keep the FormatSession test beside format.go
