+++ internal/sessioncmd/session_test.go
@@ -904,14 +904,3 @@ func TestReviveRefusesWhileTheAgentIsStillRunning(t *testing.T) {
 // FormatSession names the conversation a captured id resumes. A line that
 // omits it reads like one that names none, so an agent told to revive
 // "the conversation this session held" has to see it on the line.
-func TestFormatSessionCarriesTheCapturedConversationID(t *testing.T) {
-	withID := Session{Name: "worker", ID: "abcd1234", Tool: "claude", Directory: "/repo", AgentSessionID: "conv-42"}
-	if got := FormatSession(withID); !strings.Contains(got, "(conversation conv-42)") {
-		t.Fatalf("a captured id should name its conversation, got %q", got)
-	}
-	withoutID := withID
-	withoutID.AgentSessionID = ""
-	if got := FormatSession(withoutID); strings.Contains(got, "(conversation") {
-		t.Fatalf("a session with no captured id must not name a conversation, got %q", got)
-	}
-}
